requirements.txt
Flask==2.2.5
google-api-python-client==2.86.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-together==0.2.0
twilio==9.3.3
python-dotenv==0.21.0
requests==2.31.0
